+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+unreleased
+==========
+* Added base class for contrast coding schemes in order to make them more maintainable
+* Added hierarchical column feature in target encoder
+* Fixed maximum recursion depth bug in hashing encoder
+
 v2.5.0
 ======